src/NumSimGui/requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
numpy>=1.20.0
vtk>=9.0.0

